FacePayAPI/requirements.txt
astroid==2.3.3
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.3
fastapi==0.68.1
firebase-admin==5.0.2
google-api-core==1.31.2
google-api-python-client==2.19.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.0
google-crc32c==1.1.3
google-resumable-media==2.0.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
isort==4.3.21
keras==2.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
mtcnn==0.1.1
numpy==1.21.2
opencv-python==4.2.0.34
packaging==21.0
Pillow==8.3.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pylint==2.4.4
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.14.0
starlette==0.14.2
typed-ast==1.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
wincertstore==0.2
wrapt==1.11.2
